// server/src/firestore/dto/firestore.farmInspection.dto.ts
export type Log = {
  properties: Record<string, any>;
  isDeleted: boolean;
  approvalStatus: string;
  reviewStatus: string;
  systemStatus: any;
  isArchived: boolean;
  enabled: boolean;
  sourceSystem: any;
  meta_workspace: string;
  meta_organisation: string;
  meta_configkey: string;
  id: Record<string, any>;
  createdDate: string;
  updatedDate: string;
  lastActivityDate: any;
  createdBy: Record<string, any>;
  updatedBy: Record<string, any>;
  createdLocation: Record<string, any>;
  updatedLocation: Record<string, any>;
  parentId: Record<string, any>;
  workflowScopeId: Record<string, any>;
  auditActivityId: Record<string, any>;
  parentSecondaryItems: any[];
  allParentItemsSearchable: string[];
  primaryActivityName: string;
  primaryActivityLabel: string;
  workflowName: string;
  activityName: string;
  activityLabel: string;
  method: string;
  type: string;
  previewValues: Record<string, any>[];
  oldProperties: Record<string, any>;
};

export type InspectedFarm = {
  properties: Record<string, any>;
  surveys: any;
  surveysFull: any;
  isDeleted: boolean;
  approvalStatus: string;
  reviewStatus: string;
  reviewEntityId: any;
  systemStatus: any;
  isArchived: boolean;
  enabled: boolean;
  sourceSystem: any;
  meta_workspace: string;
  meta_organisation: string;
  meta_configkey: string;
  id: Record<string, any>;
  createdDate: string;
  updatedDate: string;
  lastActivityDate: any;
  createdBy: Record<string, any>;
  updatedBy: Record<string, any>;
  createdLocation: Record<string, any>;
  updatedLocation: Record<string, any>;
  name: string;
  contactFirstName: any;
  contactMiddleName: any;
  contactLastName: any;
  contactGender: any;
  contactDob: any;
  contactDobApproximate: any;
  contactIdentificationNumber: any;
  contactIdentificationNumberType: any;
  contactEducation: any;
  contactMaritalStatus: any;
  contactHouseHoldMemberCount: any;
  nickName: any;
  type: any;
  parentFacility: Record<string, any>;
  parentFacilityParent: any;
  parentFacilityParentParent: any;
  parentFacilityFull: Record<string, any>;
  parentLocation: any;
  parentLocationParent: any;
  parentLocationParentParent: Record<string, any>;
  parentLocationParentParentParent: Record<string, any>;
  parentLocationFull: any;
  mainContactPerson: Record<string, any>;
  mainContactPersonFull: Record<string, any>;
  location: Record<string, any>;
  address: any;
  geodatas: any;
  polygon: any;
  polygonFull: any;
  areaTotal: number;
  certificationStatus: string;
  creationStatus: string;
  isOrganic: boolean;
  complianceImplementationDeadline: any;
  numberOfPlantsTotal: number;
  areaOrganic: number;
  areaNonOrganic: number;
  areaPermanentCrop: number;
  areaAnnualCrop: number;
  cultivationStartDate: any;
  varieties: any[];
  varietiesFull: any[];
  primaryCrops: any[];
  secondaryCrops: string[];
  yieldEstimateRaw: number;
  contractDate: any;
  registrationDate: any;
  certificationStartDate: any;
  lastChemicalUseDate: string;
  lastInspectionDate: any;
  firstVisitDate: any;
  trainings: any[];
  trainingsFull: any[];
  contacts: any[];
  contactsFull: any[];
  certifications: any;
  animalCounts: any[];
  animalCountsFull: any[];
  complianceItems: any;
  plots: any[];
  plotsFull: any[];
  mobilePayWalletsFullIds: any[];
  mobilePayRegistrationStatus: string;
  identityVerificationStatus: string;
  source: string;
};

export type FirestoreFarmInspectionDto = {
  logs: Log[];
  farm?: InspectedFarm;
  entity?: InspectedFarm;
};